+++ setup.py
@@ -43,6 +43,7 @@ setup(
         ],
         "kedro.hooks": [
             "kubeflow_mlflow_hook = kedro_kubeflow.hooks:mlflow_activate_parent_run",
-        ]
+            "kubeflow_cfg_hook = kedro_kubeflow.hooks:register_templated_config_loader",
+        ],
     },
 )
